requirements.txt
chess==1.11.1
pillow==11.1.0
python-chess==1.999
